Split up Schemas into content related one and the rest, Fix the bugs and issue in the BibleVideo, Filter logic changed to int range based queries
